simplify show implementation (#31)
This PR simplifies a bit of the code used for defining show by leveraging the built-in Base functionality of replacing values for printing.
This shouldn't change the functionality, although it does change a bit of the exact alignment properties.
In particular, here, the zero-value print is used to determine things like alignment, and only afterwards is this replaced with a centered dot, while I think the previous implementation first replaced the entries with dots, and only then computed alignments.